requirements.txt
asgiref==3.7.2
distlib==0.3.7
dj-database-url==2.1.0
Django==5.0
django-heroku==0.3.1
filelock==3.12.4
gunicorn==21.2.0
packaging==24.0
pillow==10.2.0
platformdirs==3.11.0
psycopg2==2.9.9
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
virtualenv==20.24.5
whitenoise==6.6.0
